+++ gfc-client/internal/config/config.go
@@ -10,7 +10,7 @@ import (
 )
 
 const (
-	Version = "1.1.0"
+	Version = "2.0.0"
 
 	DefaultRoot      = "/opt/gfc-client"
 	DefaultEtc       = "/etc/gfc-client"